+++ tests/test_cases/dart/extra_args/input/source.dart
@@ -1,0 +1,9 @@
+class _Box {
+  final int value;
+  _Box(this.value);
+  int get doubled => value * 2;
+  int scale(int factor) {
+    var local = factor;
+    return value * local;
+  }
+}