+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add native extension fields support for Java
 - Add native Rust based Swift parser that uses WASM tree-sitter grammar
 - Add native Kotlin tag generator that uses downloadable WASM tree-sitter grammar
+- Add native Dart tag generator that uses downloadable WASM tree-sitter grammar
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers
@@ -27,6 +28,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Java plugin removed from the source repository
 - Swift plugin removed from the source repository
 - Kotlin plugin removed from the source repository
+- Dart plugin removed from the source repository
 
 - [0.18.0]